+++ actionview/lib/action_view/railtie.rb
@@ -76,7 +76,7 @@ module ActionView
     initializer "action_view.per_request_digest_cache" do |app|
       ActiveSupport.on_load(:action_view) do
         unless ActionView::Resolver.caching?
-          app.executor.to_run ActionView::Digestor::PerExecutionDigestCacheExpiry
+          app.executor.to_run Digestor::PerExecutionCacheExpiry
         end
       end
     end