+++ actionview/test/template/lookup_context_test.rb
@@ -6,7 +6,6 @@ require "abstract_controller/rendering"
 class LookupContextTest < ActiveSupport::TestCase
   def setup
     @lookup_context = ActionView::LookupContext.new(FIXTURE_LOAD_PATH, {})
-    ActionView::LookupContext::DetailsKey.clear
   end
 
   def teardown